requirements.txt
numpy==1.19.2
tqdm==4.55.1
opencv_python==4.5.2.52
PyYAML==5.4.1
